+++ common/src/main/java/dev/mayaqq/estrogen/registry/EstrogenEvents.java
@@ -2,18 +2,17 @@ package dev.mayaqq.estrogen.registry;
 
 import dev.mayaqq.estrogen.config.EstrogenConfig;
 import dev.mayaqq.estrogen.registry.effects.EstrogenEffect;
+import dev.mayaqq.estrogen.registry.recipes.inventory.EntityInteractionInventory;
 import dev.mayaqq.estrogen.utils.Time;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
+
+import java.util.concurrent.atomic.AtomicReference;
 
 import static dev.mayaqq.estrogen.registry.EstrogenAttributes.BOOB_GROWING_START_TIME;
 import static dev.mayaqq.estrogen.registry.EstrogenAttributes.BOOB_INITIAL_SIZE;
@@ -22,17 +21,14 @@ import static dev.mayaqq.estrogen.utils.Boob.boobSize;
 
 public class EstrogenEvents {
     public static InteractionResult entityInteract(Player player, Entity entity, ItemStack stack, Level level) {
-        if (entity.getType().is(EstrogenTags.Entities.URINE_GIVING) && !player.isSpectator()) {
-            if (stack.is(Items.GLASS_BOTTLE)) {
-                if ((entity instanceof Animal animal && !animal.isBaby()) || !(entity instanceof Animal)) {
-                    level.playSound(null, player.blockPosition(), SoundEvents.BOTTLE_FILL, SoundSource.PLAYERS);
-                    if (!player.isCreative()) stack.shrink(1);
-                    player.getInventory().placeItemBackInInventory(new ItemStack(EstrogenItems.HORSE_URINE_BOTTLE.get()));
-                    return InteractionResult.SUCCESS;
-                }
+        AtomicReference<InteractionResult> result = null;
+        level.getServer().getRecipeManager().getAllRecipesFor(EstrogenProcessingRecipes.ENTITY_INTERACTION.getType()).forEach(recipe -> {
+            EntityInteractionInventory inv = new EntityInteractionInventory(player, entity, stack);
+            if (recipe.matches(inv, level)) {
+                result.set(InteractionResult.SUCCESS);
             }
-        }
-        return null;
+        });
+        return result.get();
     }
 
     public static void onPlayerJoin(Entity entity) {
