+++ common/src/main/java/dev/mayaqq/estrogen/registry/EstrogenRecipeRegistries.java
@@ -10,5 +10,4 @@ import net.minecraft.world.item.crafting.RecipeType;
 public class EstrogenRecipeRegistries {
     public static final ResourcefulRegistry<RecipeSerializer<?>> RECIPE_SERIALIZERS = ResourcefulRegistries.create(BuiltInRegistries.RECIPE_SERIALIZER, Estrogen.MOD_ID);
     public static final ResourcefulRegistry<RecipeType<?>> RECIPE_TYPES = ResourcefulRegistries.create(BuiltInRegistries.RECIPE_TYPE, Estrogen.MOD_ID);
-
 }
