+++ fabric/src/main/java/dev/mayaqq/estrogen/fabric/integrations/emi/EmiCompat.java
@@ -15,7 +15,7 @@ import dev.emi.emi.api.widget.Bounds;
 import dev.mayaqq.estrogen.Estrogen;
 import dev.mayaqq.estrogen.fabric.integrations.emi.recipes.CentrifugingEmiRecipe;
 import dev.mayaqq.estrogen.registry.EstrogenBlocks;
-import dev.mayaqq.estrogen.registry.EstrogenRecipes;
+import dev.mayaqq.estrogen.registry.EstrogenProcessingRecipes;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
@@ -72,11 +72,11 @@ public class EmiCompat implements EmiPlugin {
 
         registry.addWorkstation(CENTRIFUGING, EmiStack.of(EstrogenBlocks.CENTRIFUGE.get()));
 
-        addAll(registry, EstrogenRecipes.CENTRIFUGING, CentrifugingEmiRecipe::new);
+        addAll(registry, EstrogenProcessingRecipes.CENTRIFUGING, CentrifugingEmiRecipe::new);
     }
 
     @SuppressWarnings("unchecked")
-    private <T extends Recipe<?>> void addAll(EmiRegistry registry, EstrogenRecipes type, Function<T, EmiRecipe> constructor) {
+    private <T extends Recipe<?>> void addAll(EmiRegistry registry, EstrogenProcessingRecipes type, Function<T, EmiRecipe> constructor) {
         for (T recipe : (List<T>) registry.getRecipeManager().getAllRecipesFor(type.getType())) {
             registry.addRecipe(constructor.apply(recipe));
         }
