+++ fabric/src/main/java/dev/mayaqq/estrogen/fabric/integrations/rei/ReiCompat.java
@@ -15,7 +15,7 @@ import com.simibubi.create.infrastructure.config.CRecipes;
 import dev.architectury.fluid.FluidStack;
 import dev.mayaqq.estrogen.fabric.integrations.rei.categories.CentrifugingCategory;
 import dev.mayaqq.estrogen.registry.EstrogenBlocks;
-import dev.mayaqq.estrogen.registry.EstrogenRecipes;
+import dev.mayaqq.estrogen.registry.EstrogenProcessingRecipes;
 import dev.mayaqq.estrogen.registry.recipes.CentrifugingRecipe;
 import io.github.fabricators_of_create.porting_lib.mixin.accessors.common.accessor.RecipeManagerAccessor;
 import me.shedaniel.rei.api.client.gui.Renderer;
@@ -100,7 +100,7 @@ public class ReiCompat extends CreateREI {
         ALL.clear();
 
         CreateRecipeCategory<CentrifugingRecipe> centrifuging = builder(CentrifugingRecipe.class)
-                .addTypedRecipes(EstrogenRecipes.CENTRIFUGING)
+                .addTypedRecipes(EstrogenProcessingRecipes.CENTRIFUGING)
                 .catalyst(EstrogenBlocks.CENTRIFUGE::get)
                 .itemIcon(EstrogenBlocks.CENTRIFUGE.get())
                 .emptyBackground(177, 80)
